requirements.txt
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psutil==5.5.1
pycparser==2.19
python-gnupg==0.4.4
requests==2.21.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.14.1
